controller: Support scheduling at a Pod level.
This change modifies the Stack controller logic to work at a Pod level
rather than at the Deployment level, this means that each Elasticsearch
node is created and managed individually with all that it implies.

Adds multiple fields to the StackStatus under an Elasticsearch struct
in order to save the necessary data for the Controller operation.

The Algorithm uses a count based approach to Create / Delete instances.
It supports scaling up and down the instance count but doesn't account
for Elasticsearch configuration changes in any way, or handle data in
those instances.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>