Correclty support Pod-level scheduling.
This change completely modifies the previous algorithm tho schedule pods
using a count-based approach.
The proposed algorithm uses the stack name as the prefix followed by the
workload's type (es) and last, an autogenerated random suffix. If the
prefix without the autogenerated ID exceeds the allowed maximum length
it is trimmed to fit 63 characters.

It also supports detecting when a stack configuration has changed and on
such, creates new pods for the ones that have an outdated configuration.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>