Add basic end-to-end tests
Nothing fancy, this is just a stupid bash script performing stack
creation against a pre-configured k8s cluster with the current
kubectl profile.

Probably we'll need something better at some point, but for now it
can be considered as a relatively quick iteration to have something
we can run when we want to check stack creation is not broken.