Handle pods on K8s nodes that are cordoned
This change adds an extra watch that targets the K8s nodes on a cluster
handling the case where any of them might be marked as `Unschedulable`.
In case there's any nodes in the cluster that contain Elasticsearch pods
when they're marked as `Unschedulable` or `cordoned`, an annotation is
set to any of the pods that are present, handling that annotation as a
reason to remove the pod so that they are moved away from the nodes that
are marked as `Unschedulable` or `cordoned`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>