convert from dep to go modules
For more information about Go modules, see
https://github.com/golang/go/wiki/Modules.

The ugly bit is the practical vendoring of cmd/controller-gen/main.go
from
https://raw.githubusercontent.com/kubernetes-sigs/controller-tools/v0.1.1/cmd/controller-gen/main.go,
which is a practical move because the old tooling relied on the tooling
being available under the `vendor/` folder, which is no longer the case.

Once this is merged, developers should NOT have this repository checked
out inside of a `$GOPATH`, or ensure that `GO111MODULE=on` is set. See
https://github.com/golang/go/wiki/Modules#when-do-i-get-old-behavior-vs-new-module-based-behavior
for more details.