Run e2e tests as a k8s job
The main issue with running e2e tests locally is the access to the ES endpoint
(not exposed outside the k8s cluster).
This makes it possible to run the e2e tests as a k8s job, opening the way
for adding more tests related to querying ES endpoint.