refactor main bits of elasticsearch_controller into a composable driver
it's still not a full refactor, versionStrategy should likely go away
for now, and the driver/pods.go file is a bit of a mess (well, it's the
same mess as before, but its method signature got longer, so it's not
really helping)

not all bits a driver is intended to do are solved in this PR either,
but it keeps functionality extremely close to what's in right now.

further work include addressing the above, as well as enriching the
driver and making it more readable/understandable + documentation etc.