initial support for zen2 discovery.
Based on the temporary documentation available at
https://github.com/elastic/elasticsearch/pull/34714/files.

- When a cluster is bootstrapping for the first time (i.e the first
round of nodes are being created), the `cluster.initial_master_nodes`
setting must be set. It's safe to leave this setting as garbage on
nodes because they can only be bootstrapped once.

- When master nodes are removed, we use the vote exclusion api to
exclude them from voting so we can safely reduce the quorum size en
masse.