Make the operator aware of the namespace it's running in
This is required to configure the operator to watch resources in a single namespace only.
When running within K8s, the operator can get the current namespace from the filesystem.
When not running in K8s (local dev environment), we can pass an additional flag: `--namespace=my-namespace`, which defaults to "default".