Push latest operator image when bootstraping gke
When running the operator locally with `make run`, we implicitly rely on
the "latest" docker image of the operator for init containers.
On minikube we push it to a local registry when bootstraping the
cluster.
This PR does the same with GKE: build and push the "latest" image to
GCR.