Make fs watch test resilient to extra events
This test seemed to be flaky in CI. I think it's possible that the
filesystem sends extra events (eg. 2 events when a file is created),
depending on what fs is running.

This should fix the test by checking that at least one event happened
instead of waiting for a single event.