Replace fsnotify with our own fs polling mechanism
This replace the fsnotify-based watch we had with a poll-based
watch.

We observed some inconsistencies with fsnotify (missed event, CREATE
events with no file created). Not exactly sure what happened if there is
really a bug in fsnotify, but the lack of confidence we have led us to
rewrite our own simpler disk-polling mechanism.

The new Watcher mechanism can:
- watch files in an entire directory
- be restricted to a single or several files
- ignore hidden files
- cache files content in memory, to compare new files with current cache
everytime it read the filesystem, and invoke the passed event handler
only when a file was created/updated/deleted

Both cert-initializer and keystore-updater are using it.