Fail test immediately on test.RetryUntilSuccess error
By using `require` instead of `assert`, we fail a unit test immediately
if the retry never succeeds.

This does not change the behaviour of the test itself. It avoids
situations where something fails, but we go on, provoking next tests to
also fail. In such situations, it's pretty hard to notice in CI logs
that the real failure is the first test, and next failures are just
provoked by the first one.