Address version 7.0 API deprecations (#536)
This adds separate Elasticsearch clients for version 6.x and version 7.x to be able to use the appropriate API calls for each supported version. Version selection is done based on the running pods with a fallback to the Elasticsearch spec. Supporting processes like snapshotter and keystore updater have been amended with a version flag.