Enqueue a reconciliation when a cluster health change is observed (#564)
The goal of this PR is to enqueue a reconciliation request when a cluster health changes, as seen from the observers.
This should fix inconsistencies in e2e tests where cluster keep an "unknown" state forever.

Implementation notes

To keep ES observers and k8s stuff (including reconciliation loop) loosely coupled, I introduced a way to register listeners to the observers. They get executed on every observation event.

Then, we set up one listener that checks for changes in the cluster health (comparing previous to current state). In case of change, it enqueues an event for the associated cluster.

The event is enqueued through a source.Channel (it basically emits a GenericEvent into a channel). Then, it is processed by a GenericEventHandler (also introduced here), whose job is simply to enqueue a reconciliation request for the resource corresponding to the GenericEvent Namespace and Name.


* Add a hook function in the observer and simplify unit tests by using it

* Support observation listeners to execute a func on every observation

* Add a listener looking for cluster health change on each observation

* Enqueue a reconciliation on cluster health change

* Add an integration test for the observer enqueue mechanism

* Fix variable scope bug in for loop and improve unit test

* Fix typos in comments

* Fix imports formatting