Add an example of how to deploy an nginx ingress with TLS and basic auth (#571)
* Add an example of how to deploy an nginx ingress with TLS and basic auth

The sample provides a working example (tested manually with my own
domain name) that sets up an ingress in front of both elasticsearch and
kibana.

It supports:
- TLS termination at the NGINX level (optional)
- TLS cert auto-generation from Let's Encrypt (optional)
- Basic auth at the NGINX level (optional)
- Routes rewriting
- IP filtering (whitelist)
- HTTPS communication from NGINX to ES (if using XPack security): this
is done in "insecure" mode. There is a 'secure-verify-ca-secret:
secret-name' annotation that we could use to provide a key and cert, but
it requires some extra changes in the way we store the secret.

The goal here is not to demonstrate how exposing ES *must* be done, but
a way to do it nicely.