Remove explicit snapshot support in the operator
We decided to move away from an explicit snapshot support in the
operator that could potentially give us a hard time once the stack
snapshot feature evolves.

Instead, we'll provide support for generic secure settings (including
snapshot repository credentials), and provide documentation for users to
setup their own snapshot CronJob.

This commit removes:
* snapshot repository support in the spec
* associated doc and samples
* snapshot repository secret reconciliation
* the snapshotter cronjob (and its reconciliation)
* snapshot support in the ES client implementation (still correct but
not used anyore)

However this is kept around:
* a keystore secret, mounted in the pod. For now it's an empty
placeholder secret in which we'll inject secure settings in a later
commit.
* the keystore-updater, that we'll still use to inject secure settings