Process manager lifecycle (#610)
Ties the lifecycle of the process-manager to the Elasticsearch process all the time. If one of the two dies, the second dies, even for a stop/kill perform via an API call.

The state of the process is persisted in a file /tmp/process.state to allow the process manager to decide whether to restart the process after a restart of the container.

A channel can be passed to the Start() to chose what to do when the process exits. The main reason behind this is to be able to easily test the process manager.