Implement full cluster restart (#665)
This commit adds support for full cluster restart, using the process manager in each Elasticsearch pod to restart the ES process without recreating the pod.

I hope it can also serve as a basis for implementing rolling restart for pods reuse.

The restart can be triggered by annotating the cluster resource:

annotations:
    elasticsearch.k8s.elastic.co/restart: coordinated
In turn, the operator will automatically annotate all pods that need to be restarted with a few annotations:

annotations:
    elasticsearch.k8s.elastic.co/restart-strategy: coordinated
    elasticsearch.k8s.elastic.co/restart-phase: schedule
    elasticsearch.k8s.elastic.co/restart-start-time: <timestamp>

The restart itself is then processed by a state machine, persisted through the "phase" annotation on each pods. The coordinated strategy ensures all pods move from one phase to another together (stop them all, wait for all to be stopped, start them all, wait for all to be started).
Operations done in each phase are idempotent: they are repeated over and over again at each reconciliation until all pods can move over to the next phase, and our apiserver cache is up-to-date.

Restarting an individual ES node (or a set of nodes) is also possible through the same annotations.

I included basic e2e tests in the PR.

Something important: once we have started a restart process, the rest of the reconciliation loop (pods creation, deletion) is blocked. I applied a global timeout to the restart process, so that we abort the restart and unlock the reconciliation in case something is wrong with the cluster.