Make the keystore updater start non blocking
By making the Keystore updater wait for Elasticsearch to be ready,
its start-up blocks the start of the process manager.
Until the process manager start-up is not finished, the process
manager does not start to wait to exit when Elasticsearch terminates.
As a result, if Elasticsearch dies during its start-up, the process
manager will not die and remains blocked by the Keystore updater
waiting the died Elasticsearch.