Copy ES CA into Kibana namespace during association reconciliation
This allows an association between Elasticsearch and Kibana to be
established when both live in different namespaces.

In order to do so, we copy the content of ES CA certificate into
another secret in Kibana's namespace.
Kibana CA secret is then mounted into Kibana pod.

ES CA secret is watched by the association controller, to reconcile on
any CA content change.