Create distinct es clients for observer and main reconciliation
If we use the same client that gets closed at the end of a
reconciliation, it would otherwise impact observers' client.

Note: most of the times, the client passed to the observer is not used.
Observer keeps using the same client it was created with unless the new
one is different (different certs, auth, etc.).