Turn most of dynamic watches integration tests into unit tests
This commit sets up a unit test to cover the integration test scenario
we had for the DynamicEnqueueRequest handler. A very basic integration
test is left to ensure everything correctly fits together, but most of
the logic we wanted to test (and a bit more) is moved into the unit test.

Main benefits:
* remove complexity from IT complexity (edge cases of 1-2 potential reconciliations,
wait for something to not happen)
* remove dependency to test.Timeout in most tests (used both in
`CheckReconcileCalledIn` and `CheckReconcileNotCalledWithin`)
* faster execution time