Build ES pod on top of user-provided podTemplate
Make the elasticsearch pod and container inherit user-provided
podTemplate from the Elasticsearch spec.

It allows the user to pass-in custom annotations, labels, init
containers, environment variables, volumes, volume mounts, etc.

We append our own spec on top of the user's one, or set defaults in case
the user has not provided a particular field.

When dealing with list of items (eg. volumes, volume mounts,
environment), this commit does not attempt to deal with conflicts,
but only append our own items to the user-provided ones.
In practice, that means:
- setting a volume mount path conflicting with ours will return an error
at pod creation time
- setting an environment var conflicting with ours will lead to the pod
having both env vars defined in the spec, but the final container env
will only see the last-defined variable (ours, not the user's one)