Build Kibana Deployment pod template on top of user-provided podTemplate
This commit changes the way we build the Kibana pod template to inherit
from a user-provided template.
This allows the user to set their own labels, annotations, environment
variables, volumes, etc.

Note that it does not attempt to deal with conflicting items
(environment variables, volumes, etc.). To be done in a follow-up PR.

It also does not deal with improving (fixing) the Kibana spec comparison
mechanism: to be handled with
https://github.com/elastic/cloud-on-k8s/issues/772.

It also includes some minor refactoring:
* move labels into their own package to avoid circular imports
* get rid of DeploymentParams that just propagates more and more fields
from the Kibana spec: just use the spec directly
* add a helper function to retrieve the Kibana container from the spec,
instead of operating on all containers of that spec