Don't override user-provided labels (#957)
We decided we should keep whatever labels users apply in the podTemplate
section of the spec. Even though this may lead to issues for conflicting
labels with the ones we set. Label names are very explicit so this kind
of conflict probably originates from users desire, we should not work
against that.