Remove auto-generated CRD integration tests (#955)
I think those tests are a bit useless for us, they are testing the
underlying kubebuilder code generation more than they are testing
anything the operator brings.

Plus they sometimes fail for various reasons and we don't use the same
rety mechanism we use for other integration tests relying on a local
apiserver. We could use it but it needs a slightly different plumbing.
Probably not worth it considering the above.