Remove the cert-initializer.
The workflow was more complex and gained very little. Certificates and keys for
the transport protocol for ES Nodes are now genereated by the operator and delivered
via a mounted secret.

To prevent ES from attempting to start before the certificates have been delivered
to the node, a small wait-until-tls.key-exists-snippet has been added to the prepare-fs
init container.