Create a default PDB for ES Pods.
To configure or disable the default PDB:

```yaml
apiVersion: elasticsearch.k8s.elastic.co/v1alpha1
kind: Elasticsearch
metadata:
  name: elasticsearch-sample
spec:
  version: "7.1.0"
  nodes:
  - nodeCount: 5

  ## to opt out of the default PDB, set it to the empty object
  #podDisruptionBudget: {}

  # to configure the defaulted PDB:
  podDisruptionBudget:
    # custom metadata for the budget, metadata.name and metadata.namespace will be ignored
    metadata:
      labels:
        foo: bar

    # customize the PDB Spec (see v1beta1.PodDisruptionBudgetSpec):
    spec:
      # default maxUnavailable is 1
      maxUnavailable: 2
```

Make podDisruptionBudget a top level field in the ES spec

Removes the `podDisruptions` wrapper and changes the sematics to disable the
default if an empty `podDisruptions` object is provided (e.g `{}` in YAML). This
gets rid of the `disableDefaultPodDisruptionBudget` field since the semantics above
allows for disabling as well.