Remove JVM DNS cache policy tweaks
Since we only support Elasticsearch 6.8+ and 7.1+, we have the following
commit already there in Elasticsearch:
https://github.com/elastic/elasticsearch/pull/36570.

Hence we do not need to duplicate the fix manually into our own JVM
security options file.

Since there is now no need for this "default config map" volume in the
pod (only used for the jvm security options file so far), this commit
also removes it.