Improve pod-pvc labelling and comparison (#1129)
* Improve pod-pvc labels comparison

By introducing a hash of the pod template in the pod, and copying this
over as part of all pods labels into created PVCs, we accidentally tied
PVCs to be reused only for pods that match the exact template the
initial pod was created with. This prevents reusing PVCs, for example,
when replacing a 2GB pod by a 4GB pod.

This fixes it, by restricting labels comparison to a subset of all pod
labels. It also compares ES version to make sure we don't give data from
a newer version to a pod running an older version of Elasticsearch.
Labels copied over to the PVC from the pod are now also limited to a
subset of all pod labels, not including the pod template hash.

* Don't copy pod annotations over to pvcs

This seems quite useless, and can be confusing.