Respect volume name when reusing PVCs (#1122)
Puts the volume name into the PVCs labels to include it in the comparison.
Enables the previously commented e2e that tests the default case of reuse.
Adds another e2e test to make sure PVCs are reused while respecting their name.

I ran into an issue with scheduling pods with more than one PVC, because we use volumeBindingMode: Immediate more often than not the volumes would be allocated in two different zones which meant that the pod became unschedulable as it cannot be in two zones at once.

The solution I found was to switch to volumeBindingMode: WaitForFirstConsumer which creates the volumes only once the pod has been scheduled to a node. The only problem with this is that in k8s 1.11 this does not work because volumes need to be pre-provisioned. In 1.12 DynamicProvisioningScheduling is no longer feature gated and volume provisioning with late binding works as expected. So this PR also

* adds a storage class template with WaitForFirstConsumer
* ups the default k8s version to 1.12
* skips the new e2e test on 1.11