Set a default memory resources request for Elasticsearch
We recently disabled any default resources: as a result, it is possible
to deploy eg. 30 Elasticsearch nodes on a 1GB machine. They will all
randomly crash (OOM), which is quite confusing.

This commit enables a default memory resources request of 2Gi, matching
the default ES heap size of 1Gi, if the user does not provide their own
resource requirements.

For cases where the user wants to rely on the k8s cluster configured
LimitRanges, hence not provide resource requirements in the spec, it is
possible to explicitly pass empty resource requirements in the spec:

podTemplate:
  spec:
    containers:
    - name: elasticsearch
      resources:
        limits: {}

In this situation, the operator will not apply default resources.