Reconcile on statefulsets changes & pod changes
Pods are not owned by our controller anymore, but by the StatefulSet
controller. We rely on pod statuses updates to reconcile certificates,
hence we need to trigger reconciliation on any pod update.

Let's setup a watch for pods who belong to an Elasticsearch cluster.

Also, let's make sure we watch StatefulSets for changes.