Handle sset scale up/down & replacement
Add support for scaling a StatefulSet up (just update sset replicas) or
down (migrate data away then remove nodes when ready). Which also adds
support for renaming a StatefulSet, by creating the new sset, then
slowly remove the existing one with data migration.

It ignores any zen1/zen2 consideration for now (works fine with zen2 in most
cases).
It ignores any change budget consideration for now.
Unit tests and E2E tests are missing, but a refactoring is to be
expected to handle zen2/zen2 and the changeBudget. I'd prefer waiting
for this refactoring to happen before dealing with large unit testing.
Consider this as still work in progress.