Replace Pod-based reconciliation by StatefulSet-based reconciliation (#1205)
Build a set of StatefulSets matching the ES CRD NodeSpec, and reconcile them instead of reconciling pods.

This is a very first iteration.

Bootstrapping an ES cluster seems to work fine. Upgrading this cluster (and everything related to zen1/zen2) does not.

Main bits of this change:

- Require the name field in the NodeSpec, to derive the StatefulSet name
- Build a StatefulSet per NodeSpec, along with a headless service
- Label the StatefulSet with a hash of its template, for comparison purpose
- Create a single Config secret per StatefulSet (instead of per pod)
- Label the pod template with a hash of the config (before zen2 initial master nodes), to rotate pods on config changes
- Reconcile StatefulSet, service and config in the default driver, commenting out most of the existing pod reconciliation logic

The code is left in a rather dirty state: I had to make a few changes deep down in functions that build pod spec and config. These changes bubble up pretty high, to changes calculations and the default driver.
Most of the default driver code and the mutation package is probably doomed, I chose to keep it there commented instead of deleting it for now, since it I think it may help in the short-term refactoring ("where does that logic live in the new code? here? OK I can remove."). Let me know if it still deserves more cleanup.

To be refactored and improved in follow-up commits.